Feat: Custom Endpoint - specifying Date ranges
What?
(1) we can now call custom date ranges for the events we want to display in our backend - not sure how this will be implemented for Google Calendar (figure out later)
(2) now we are able to send a get request for a specific date range and it will return events within that date range.
 this will be super useful for our front end to display onyl a specific number of events